requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
aiosqlite